+++ ChestsAnywhere/docs/release-notes.md
@@ -1,6 +1,11 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* In the category dropdown...
+  * Cellars are now translated and include the name of the player who owns them.
+  * Building interiors now use the translated name for the building, if available.
+
 ## 1.28.3
 Released 27 July 2025 for SMAPI 4.1.10 or later.